+++ src/main/java/com/global/api/entities/enums/SuspiciousAccountActivity.java
@@ -1,0 +1,6 @@
+package com.global.api.entities.enums;
+
+public enum SuspiciousAccountActivity {
+    SUSPICIOUS_ACTIVITY,
+    NO_SUSPICIOUS_ACTIVITY
+}